3_container_send_events/requirements.txt
azure-eventhub
azure-eventhub-checkpointstoreblob-aio
pyodbc
flask
flask-restful
